Use convert to create jpg fallbacks, updates for Inkscape 1.0
Inkscape 1.0 does not support native jpg exports anymore. Instead, they
recommend to export to PNG and use convert (from imagemagick) to create
the jpg file from that.

As such, we introduce options to configure the convert program with
default options and use it.

For Inkscape 1.0, we remove the deprecated --without-gui argument. This
is the default now.